initial v5.0

+++ a_models/smpl.py
@@ -29,19 +29,19 @@ class SMPL(pl.LightningModule):
         else:
             self.faces = None
 
-        np_v_template = np.array(model['v_template'], dtype=np.float)
+        np_v_template = np.array(model['v_template'], dtype=float)
         self.register_buffer('v_template', torch.from_numpy(np_v_template).float())
         self.size = [np_v_template.shape[0], 3]
 
-        np_shapedirs = np.array(model['shapedirs'], dtype=np.float)
+        np_shapedirs = np.array(model['shapedirs'], dtype=float)
         self.num_betas = np_shapedirs.shape[-1]
         np_shapedirs = np.reshape(np_shapedirs, [-1, self.num_betas]).T
         self.register_buffer('shapedirs', torch.from_numpy(np_shapedirs).float())
 
-        np_J_regressor = np.array(model['J_regressor'], dtype=np.float)
+        np_J_regressor = np.array(model['J_regressor'], dtype=float)
         self.register_buffer('J_regressor', torch.from_numpy(np_J_regressor).float())
 
-        np_posedirs = np.array(model['posedirs'], dtype=np.float)
+        np_posedirs = np.array(model['posedirs'], dtype=float)
         num_pose_basis = np_posedirs.shape[-1]
         np_posedirs = np.reshape(np_posedirs, [-1, num_pose_basis]).T
         self.register_buffer('posedirs', torch.from_numpy(np_posedirs).float())
@@ -49,7 +49,7 @@ class SMPL(pl.LightningModule):
         self.parents = np.array(model['kintree_table'])[0].astype(np.int32)
 
 
-        np_weights = np.array(model['weights'], dtype=np.float)
+        np_weights = np.array(model['weights'], dtype=float)
 
         vertex_count = np_weights.shape[0]
         vertex_component = np_weights.shape[1]
@@ -143,10 +143,10 @@ if __name__ == '__main__':
         2.04248756e-01, -6.33800551e-02, -5.50178960e-02,
         -1.00920045e+00, 2.39532292e-01, 3.62904727e-01,
         -3.38783532e-01, 9.40650925e-02, -8.44506770e-02,
-        3.55101633e-03, -2.68924050e-02, 4.93676625e-02], dtype=np.float)
+        3.55101633e-03, -2.68924050e-02, 4.93676625e-02], dtype=float)
 
     beta = np.array([-0.25349993, 0.25009069, 0.21440795, 0.78280628, 0.08625954,
-                     0.28128183, 0.06626327, -0.26495767, 0.09009246, 0.06537955])
+                     0.28128183, 0.06626327, -0.26495767, 0.09009246, 0.06537955],dtype=float)
 
     vbeta = torch.tensor(np.array([beta])).float().to(device)
     vpose = torch.tensor(np.array([pose])).float().to(device)
